Reorganizing to show the core of OpenRelativity could be exported as a unity package and separate the example
Thought how to handle movement is inside the example, but at the same time is not ready for share as a library ready
to use, there is also no short description of how to setup from other game using it (i.e. the button names or the
layers). This could be improved.