Debugging collision
-Real space to Minkowski space transforms should virtually always put
the player at the origin of both spaces
-To transform a direction vector to Minskowski space, it probably makes
more sense to contract the direction vector as as if a distance and then
normalize it.
-Collision result velocites should naturally never exceed the speed of
light if the physics is handled correctly, even without a velocity
check, but they could still exceed the max velocity, so a check was
added.
-(NOTE: There seems to a be an asymmetry in length contraction depending
on whether the player is looking toward or away from the world origin.
It's visible inside the sphere. This shouldn't happen, but I can't find
what the problem is, yet.)