Attempting to correct asymmetry in length contraction
When RelativisticObject velocities change, their Minkowksi space
position should be invariant. The player is always at the effective
Minkowski space origin, but the Unity world coordinates origin is not,
then. To keep in line with Unity world coordinates, change in length
contraction between the player and world origin needs to be accounted
for when the player velocity changes.

This is an attempt to correct for this. (Might kinda still feel screwy
in the sphere in practice, but that might be due to relativistic
velocity addition and paths that aren't exactly radial in the sphere.)