Faster OpticalToWorldSearch, turned off drag
Commented out high level OpticalToWorldSearch for reference, and
replaced it with a lower level version that should at least be lighter
on allocation. Also turned off drag, because it is expensive and not yet
correctly accounting for changes in optical time due to changes in
player position. (It can be turned on easily to test
OpticalToWorldSearch.)