Corrected acceleration calculation (without Newtonian gravity)
Acceleration that was being used was roughly proportional to proper
acceleration, but was scaled to minimal effect. Trial and error
suggesting flipping gtt to gtt^-1 in usage. Had to tune MovementScript
acceleration scales to make results sensible. (These results should be
considered tentative, but the developer is considering that player
acceleration actually has a large qualitative effect at these velocity,
acceleration, and speed of light scales, hence the simulation should be
expected to be qualitatively different. There is also an occassional
"hiccup" when accelerating from negative to positive player world
velocity, or vice versa.)