Tuning physical parameters
Tuning control acceleration and base speed of light. (Note that, if the
change to an accelerated frame appeared to propagate out from the player
with infinite velocity, the apparent "change" would reach distant
objects just in time to change their apparent optical behavior
immediately from the player's perspective, as is implemented. This can
be checked assuming a speed of light propagation delay and infinite
velocity of the "propagating" change in the metric due to acceleration.)