Corrected WorldToOptical and inverse
WorldToOptical and OpticalToWorldSearch did not check for the zero
vector (or close to it) when constructing a rotation quaternion. This
led to a noticable inconsistency between the shader and the utility
methods in the edge case. (It might be preferable to rather eliminate
this check, but the utility is now at least consistent with the shader.)