Optimization
UpdateCollider was instantiating more often then necessary. The choice
of updating physics in "Update()" rather than "FixedUpdate()" was
probably by design, but it's unclear whether it provided any benefit.
Also, garbage collection and reallocation can be avoided much of the
time by only reallocating vertex arrays when the "rawVerts" array is too
small to hold the current mesh vertices array. (The length of the actual
coding vertices array is stored as a separate variable. As can be seen
by how it used for the compute shader, an overallocated vertex buffer
probably does not generally pose a problem for use of "rawVerts.")