Cheap colliders
(Collision reaction still seems too energetic.) Depending on the
platform, editing mesh collider verts can be prohibitively expensive to
do every frame. An alternative might be to subdivide box colliders, like
with ObjectMeshDensity, and Lorentz transform their positions and
scales. Rather than lay the entirity of the work onto the CPU, a compute
shader can be used for GPGPU transformation purposes. The result is less
accurate than mesh collider vert transformation, but also much less
expensive. (Hard to profile, but the difference in overhead appears to
come primarily from mesh re-baking for physics.)

Also, it doesn't seem right that drawing should be turned on and off in
the shader on a per-vertex basis. (It's also additional shader overhead,
with alpha.) This task should probably be given to the GameObject, and
operate on a per-object basis.