Relativistic lighting (see description)
(Long description needed. These features are still experimental:)

Relativistic Doppler shift from a moving mirror is a well-known problem
within the physics community. The reflective Doppler factor goes like
the square of the emissive doppler factor, for the respective
appropriate velocities. Heuristically, this can be conceptualized as due
to the apparent velocity of the image of the emitting source behind the
(specular) reflecting surface. (
https://arxiv.org/pdf/physics/0605100.pdf ) We assume that diffuse
reflection exhibits the same Doppler factor as specular reflection.

Take an inertial lab rest frame, an emitter moving relative the lab
frame, and a reflector moving relative to both the emitter and the lab
frame. Say the emitter emits red light in its rest frame, which appears
blue in the lab frame at the point of closest approach to the reflector.
In the lab frame, the blue-shifted light from the red emitter is
indistinguishable from blue light of the same frequency emitted by an
object at rest with respect to the lab frame, in terms of the
indistinguishability of photons of the same frequency. This is the key
physical fact that allows relativistic lighting without drastic redesign
of Unity's lighting system, that Doppler-shifted light from a moving
emitter, relative the Unity world coordinate rest frame, is physically
indistinguishable from light of the same frequency emitted by an
observer that is stationary with respect to Unity world coordinates.

Reflected light carries a squared Doppler factor, compared to emitted.
We book-keep ambient light by its frequency at closest approach with
respect to the Unity world coordinate rest frame. Whatever the velocity
of the emitter, we Doppler shift light to the world frame. Reflectors
(specular or diffuse) can treat this frequency in the world frame as if
it were emitted by a stationary observer in the world frame. (There is
no static propagation medium to consider, unlike sound.) Then, it's
simple to apply the correct respective emissive and reflective Doppler
shifts, (the original shader's shift and its square).

Again, this is experimental, and it might be subject to change. The
theory and application is still new to the developer.