Previous commit not perfectly right, but acceptable for static
The previous commit explanation is not exactly right. In the case of
moving light sources, (including emissive materials,) the given
frequency of emission in the world frame depends on the angle between
reflector and camera.

However, this works for static light sources and emissive materials,
with respect to world coordinates, which are a common use case. We bake
static materials, anyway. The meta pass on the emissive materials has
been cut down to take advantage of this fact and make baking more
efficient. However, forward pass won't be right for sources of light
moving relative the world frame.