Optimizing shaders
Thinking about it, would be feasible to correct the lighting forward
pass only for moving light sources with constant velocity. If emitter
velocities change, the speed of light delay off reflectors, to the
player, becomes extremely complicated, and depends on the history in a
way that isn't easily tracked. However, baked static lighting is often
preferred for performance reasons anyway, and the current code seems to
achieve this.