Changed unlit shader from emission to reflectance Doppler shift
For consistency when using unlit and lit shaders at the same time, the
unlit shader should have its Doppler shift squared.

The "XYZ" values appear to a spectral intensity curve by frequency.
That's why the "XYZ" values are shifted by the third power of the
Doppler factor, like intensity by frequency. However, this is debatable,
but the shader generally simulates a diffuse reflectance spectrum, not a
black body or spectral emission spectrum. If I understand "XYZ"
correctly, the Doppler factor should be squared.

(On a personal note, sorry to be "that guy," MIT Game Lab. Your physics
module is awesome, and my "Tachyoid" superset module, which I'm using
for my own game, releasing the relevant parts to this repo, wouldn't be
worth anything without the original OpenRelativity project. Thank you!)