Fixing vertex point/spot lights
RelativisticObject initialization and updating was slightly modified so
that a RelativisticObject could be attached to a spot or point light,
with "Nonrelativistic Shader" option enabled, to keep point lights at
the correct Lorentz-transformed point in space without further Lorentz
transform in the shader.

Shader quantities for vertex lights were debugged. Shader does not get
correct direction from spotlights, yet, but this will be corrected soon.