Point lights
Adding point lights and debugging vertex lights.

(Can't achieve parity between normal vector behavior for vertex and
point lights in a test scene, so currently treating normal as if it
reflects from both sides of mesh. This is a kludge, but may be either
desirable or undesirable depending on the use case.

Hopefully can add a test scene by the weekend, time permitting.)