Conformal factor
First (trivial) conformal factor example. Have not yet corrected shader
for nontrivial conformal factor.

General relativity assumes that space is locally flat, over small enough
regions. That is, over a small enough region, the metric appears to be
locally Minkowski. We can calculate the effects due to player
acceleration, as if space were flat, and then multiply the accelerated
flat metric by a conformal factor to make the space appear generally
curved. However, this also changes the paths of light rays, ("null
geodesics,") and the shaders still haven't been modified to account for
this.

(Without the stress-energy-momentum distribution causing the curvature
of the metric, according to the Einstein field equations, this isn't
general relativity, yet. However, we are taking steps toward  special
relativity, with accelerated frames, on a generally curved background
spacetime.)