Rough "Black Hole" Demo
The demo is still rough, but here's a conformal map object for a
(static, stationary throughout time) Schwarzschild metric. I don't have
every detail worked out, like I need to carry a factor of the speed of
light squared somewhere either in the metric or the RelativisticObject
coordinates, but I was excited to get this up on the repo.