Schwarzschild gravity demo
Effects outside the event horizon of the "black hole" were not very
drastic. However, this is because the speed of light is so slow! Holding
the strength of gravity fixed, but reducing the speed of light to 40
m/s, Earth's Schwarzschild radius would be roughly 5*10^11 m, compared
to its planetary radius of 6(.371)*10^6 m, roughly 5 orders of magnitude
smaller! With the normal earth surface gravity acceleration, objects on
the surface of Earth would be inside a massive event horizon! So the
gravitational effects are only realy noticeable inside the event horizon
in the first place, for an applied Schwarzschild conformal factor. The
demo scene demonstrates gravity in the interior of a black hole, around
the solid "planet" in the middle of the room.

I will attempt to incorporate gravitational lensing, vetrex-wise and/or
pixel-wise, over the weekend. Note that the applied Schwarzschild
gravity field causes the moving cubes to fall toward the "planet"
without the application of forces to the rigidbodies and/or changing the
viw in world coordinates (before objects collide). This is because the
curved paths of the boxes are actually "straight lines" of free-fall in
a curved spacetime geometry.