Refactoring for new generalized geometry
Utility functions and shader will need to account for new, specific
WorldToOptical and OpticalToWorld functions, for each conformal factor.
This also seems to be the most natural and direct way to handle the
different geodesic equations for any given applied background metric. It
will not be possible to overlay these conformal objects, at first. The
shaders will be specific to the single background metric. This will be
more precise, but the "Nonrelativistic Shader" option could be leveraged
to more easily accommodate piece-wise or composed background metrics.

(Note: order of operations might or might not be correct in applying
uniform acceleration and background metric in order to form overall
metric, but I think it will work once WorldToOptical/OpticalToWorld are
generalized. I will test this once I have these methods implemented for
Schwarzschild.)