Clean-up and four vectors
Trying to base the entire treatment around four vectors, where
applicable, (three spatial components and one corresponding to time,
these are a natural way to express vectors in relativity).

Notice that the box closest to the player on the left stays still, with
viw = 0. This isn't correct, (maybe counterintuitively). An object with
no "peculiar velocity" should co-move with the Schwarzschild
coordinates, falling into the center of the gravity well even though its
"viw" is zero. Correcting this is top priority.