Correcting unity_ObjectToWorld bug
For basically the entirety of this fork, the unity_ObjectToWorld
transform was incorrectly implemented. There's an additional step in
that transformation to transform between object coordinates and world
coordinates. This did not become apparent until we could remove the
speed==0 check in the shader.

This was a critical bug. Schwarzschild demo still not totally working,
but will continue with Schwarzschild (and "comoving coordinates") in the
morning.