Bug fixes and partial work on black hole
Several small bug fixes. Also work on Lemaitre coordinates.

The black hole metric needs a conversion from rho to r and then from
spherical to Cartesian, at the end. (This could be accomplished with one
Jacobian.) The player should also respond to the background metric, like
by transforming the Unity world origin position relative to the player
and adding the difference to the player's position, when stationary.
Still not right.