Schwarzschild debugging
When using the Schwarzschild background, or any background besides (or
trivially including) Minkowski, we should assume that the input
coordinates _are_ co-moving coordinates, because, with zero viw, the
object should automatically co-move.

For Schwarzschild, we can assume that the input coordinates are
(co-moving) Lemaitre coordinates, and that the output position is in the
usual coordinates of the Schwarzschild solution, at least if our camera
perspective is far from the black hole. However, WorldToOptical still
doesn't keep the object tracked to co-moving position, yet.

(We're incrementally committing, here, to avoid losing good work, before
we try to change WorldToOptical or correct in a different manner.)