Simple comotion solution
Properly, the comotion is the result of the geodesic equations for a
RelativisticObject, and depends upon the Christoffel symbols.

We could pass the Christoffel symbols to the shader. However, as it
stands, we only pass the value of the metric at the (unextended,
"world") position of the transform. If we only passed the Christoffel
symbols for the same point, no accuracy could be gained on a per-vertex
basis by solving the geodesic equations per vertex.

It's cheaper, and smooth, just to factor comoving geodesic motion as a
kind of "pseudovelocity" passed to the native physics engine rigidbody,
unless we want per-vertex metric and Christoffel symbol values. The
"pseudovelocity" only time dilates and length contracts as the world
coordinates do with respect to the player. This is a fictional velocity,
and has no true physical significance as a velocity. However, it's
convenient to "keep book" this way, and we offload part of the problem
to the native physics engine.