Replacing "pseudo-velocity" with Christoffel symbols
The "pseudo-velocity" approach gets the mapping from comoving to optical
coordinates wrong. (This was checked by attempting to form an orbit that
is close to stable, which failed.) The best approach is to simply use
the Christoffel symbols and the geodesic equations.

For Schwarzschild, there should be a correction to the Christoffels for
the fact that the player is close to the black hole and moving, but we
ignore this at first.