Geodesic equations
Not necessary to pass full Christoffel symbol set to shader. Christoffel
symbols and synchronous-to-world coordinate derivatives give a four
acceleration that satisfies the geodesic equation. We integrate this
acceleration over the tisw to get the world coordinate position change
due to the geodesic equation.

(Schwarzschild metric with this factored in broke the walls.) Only
incorporated this into utility methods, and therefore only the
nonrelativistic shader option. Will incorporate this four acceleration
into the shaders, next.