Correcting Schwarzschild
The (1-rs/r) factor for the player position should have an inverse
effect on the local appearance of curvature, compared to the equivalent
factor for RelativisticObject position. "Fiddling" with
Gullstrand–Painlevé coordinates suggests the given correction, but this
might be actually ad hoc and wrong.

Geodesic acceleration is wrong, and was commented out.