Acceleration and Newtonian gravity
Sign of gravity contribution to player acceleration vector was wrong
(even though the player fall direction was correct). Even though
acceleration is not relative in the same sense as velocity. player
acceleration also contributes to the apparent position of on object
transformed by the shaders. RelativisticObject proper acceleration could
also be useful.