Correcting acceleration
Acceleration is not added linearly to velocity, but we were adding it as
if it were nonrelativistic. It's numerically difficult to find the exact
distance traversed with relativistic acceleration, but we approximate it
with the mean of initial and final relativistic velocities. Also, added
a couple of bug fixes.

(Still tinkering with player acceleration, if it's necessary, but
committing now to avoid trashing good work.)