General improvements
WorldToOptical and OpticalToWorld methods were doing unnecessary things,
like calculating the relative velocity between object and player.
Cleaned up these functions, (in addition to debugging commented-out
player acceleration code, which doesn't seem to be necessary because of
lack of a time delay between player and self).

Optimized OpticalToWorld, but not sure if this is safe, yet: "tisw" is
simply the distance between the optical position and player position
divided by the speed of light. Should work if player perspective is
correct for general ConformalMap objects. (That is, if the optical
positions are calculated correctly, everything in the space should
"appear to the player as the inverse metric," I think.)