Forking accelerated shaders from inertial shaders
Shader acceleration handling was incorrect (and still is). It will
probably be necessary to resort to numerical solutions to the hyperbolic
equations of motion for shaders with acceleration. Hence, these should
be forked from the closed form inertial cases, which will be much
cheaper for use with static and other inertial meshes.

The developer feels this is motivated because:
-Player acceleration is already on a relativistic scale, and typical
accelerations will generally be in the relativistic range.
-The developer already has a case where acceleration in the shaders is
necessary, and will be implementing it anyway.