Bringing all shaders into line for acceleration
Moved new acceleration code into all shaders.

In (non-exhaustive) tests, the inner product (x.x) only appeared to be
negative for points behind the player, opposite the direction of
acceleration, which is expected for the Unruh effect, but visual
inspection suggests something is wrong in front, as well.

To have the code usable, and to test whether the rest of it is
qualitatively correct, I commented out the Lorentz boost of the metric
back to the world frame. It is calculated in the player's rest frame,
but not boosted back before application. I know that the Minkowski
metric is invariant under Lorentz transformation, but a general metric
with off-diagonal components is only guaranteed to be covariant, as far
as I know. It might solve the problem to boost all relevant vectors to
the player frame, but I don't think it should make a difference, if I'm
understanding the transformations correctly.

Some of what I'm seeing genuinely appears to be the Unruh effect.
However, with "playable" acceleration and velocity parameters, the
visuals just "bug-out" with a Lorentz transform of the metric back to
the world frame.