Fixing inconsistent "vpc" signs
Sign of "vpc" was inconsistent in various versions of the
world-to-optical transformation. I think the sign should be flipped for
my use of it, because the vpc vector is flipped when fed into the
shader, but using the player velocity directly as "viw" instead of "vpc"
works, without an additional sign flip, from a third person perspective.