Correcting RelativisticObject four-acceleration
Player acceleration is most conveniently passed around as a proper
acceleration. For RelatvisticObject instances, the geodesic equation for
a ConformalMap might result in a non-proper four-acceleration, so it
makes sense to give the four-acceleration in world coordinates. (I
assumed additivity of four-accelerations in the same frame, but I need
to double check this.)