"Gravity Wave" Demo
I turned up the LOD on the meshes, (using the ObjectMeshDensity script
provided with the original repo,) to show something really peculiar and
cool in high detail.

When the player accelerates, or jerks, a wave-like perturbation appears
to travel through the walls of the room, due to the time-varying effect
of first-person acceleration on the metric. If I understand what's going
on, we seem to have a gravity wave, (only visible in an accelerated or
jerking frame, not existing in any inertial frame).

We have all the ingredients for it: we have special relativity, Einstein
equivalence principle, and the first-person effect of acceleration on
the apparent background metric. (We can even apply general background
curvature.) For full general relativity, we only lack coupling between
stress-energy-momentum and background curvature, (according to the
Einstein field equations,) which would cause the acceleration of the
player to produce gravity waves which could be observed in the inertial
rest frame of the room, via "frame-dragging" or some configuration of
accelerating masses.

I'd love to hear the original repo owners' opinion on this phenomenon,
if the opportunity ever arises. We ignore coupling between
stress-energy-momentum and the background metric, (limit of very low
Newton's gravitational constant, with a slow speed of light,) but I
think we've come quite a way!