Gravitational lensing for Schwarzschild (#17)
* Gravitational lensing for Schwarzschild

* Deflection might have been backwards

* Revert last commit, and nothing inside of Einstein ring (unless in front, TODO)

* Cleanup (still TODO: objects between black hole and player)

* Turn off lens when behind player

* Factor angle from cam. forward to ray to lens (but this is probably not an accurate handling)

* Effect is inversely proportional to distance from lens

* Debugging

* Turn test box back on

* Debug; better demo

* Debugging

* Debugging

* Debugging behind (though this abuses the approximation)

* Planar correction was qualitatively backwards

* Renaming shader

* Minimum impact parameter = Schwarzschild radius? ("Playing around")

* Debugging

* Experimenting with black hole round trips

Maybe: if deflection angle is greater Pi/2 < \phi <  Pi, this implies deflection from the direction of the player camera plane. For Pi < \phi < 3 * Pi / 2, this implies a decaying (exiting) orbit one revolution around the black hole, and so on. (At least, this should be true if the impact parameter is greater than the Schwarzschild radius, which it can be.)

* Hypothetical "gravity mirror"

We normally consider the gravity lens equation within a deflection angle of -pi / 2 to pi / 2. Outside of this range, there is solid angle in the point-of-view to fill, and the impact parameter outside of this range is not required to be less than the Schwarzschild radius, (indicating decaying orbit of light, in-falling). This suggest decaying orbits out-falling, successive tangent ranges of width pi incrementally adding half-orbits around the black hole. Full integer orbits are simple enough, but n + 1/2 orbits would act like a "gravity mirror," bending light from behind the player halfway around the black hole and back to the player's eyes. (Incidentally, because of the variability in distance of orbit, I think time from emission to destination still appears the same across the various periods.)

* Fixing "wobble"

* Fix reversal after dissolve

* Spherical skybox approximation

* Revert "Spherical skybox approximation"

This reverts commit 2d11ed62dff91554cca2781c4249e1ddc4f3a9b0.

* Namespace; prefabs

* Gravity lens compatibility tweaks

* Force mirror ManualUpdate in time

* Compatibility debugging

* Fixing general camera orientation

* Trying to fix projection artifacts

* Safety checks

* Gravity lens shader camera scale (to reduce artifacts from off-screen lensing)

* Flip sign in gravity lens shader

* Fixing event horizon

* Cleanup

* Deflection angle sign was correct, but block event horizon

* Schwarzschild Doppler shift